backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.7.4
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
